backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.28.0
python-multipart==0.0.9
rembg>=2.0.62
pillow>=9.5.0
httpx==0.27.0